added _check_waypoints estimating path length (t) based on waypoints and speed limits
